[feat] Add sorting for metrics sortOptions
